Cleaning up store
