+++ src/main/webapp/resources/js/pages/projects/samples-share/components/ShareProjects.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useGetProjectsManagedByUserQuery } from "../../../../apis/projects/projects";
 import { IconArrowRight } from "../../../../components/icons/Icons";
-import { setDestinationProject, setNextStep } from "../services/shareReducer";
+import { setDestinationProject, setNextStep } from "../services/shareSlice";
 
 /**
  * React component to select the destination project.
