+++ src/main/webapp/resources/js/pages/projects/samples-share/index.js
@@ -18,6 +18,8 @@ function ShareSamples({ projectId }) {
     (state) => state.reducer
   );
 
+  console.log(step);
+
   React.useEffect(() => {
     if (
       !sharedProjectId ||
