+++ src/main/webapp/resources/js/pages/projects/samples-share/store.js
@@ -2,14 +2,14 @@ import { configureStore } from "@reduxjs/toolkit";
 import { fieldsApi } from "../../../apis/metadata/field";
 import { projectApi } from "../../../apis/projects/project";
 import { projectsApi } from "../../../apis/projects/projects";
-import shareReducer from "./services/shareReducer";
+import { shareSlice } from "./services/shareSlice";
 
 export default configureStore({
   reducer: {
     [projectApi.reducerPath]: projectApi.reducer,
     [projectsApi.reducerPath]: projectsApi.reducer,
     [fieldsApi.reducerPath]: fieldsApi.reducer,
-    reducer: shareReducer,
+    reducer: shareSlice.reducer,
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware().concat(
